Rename initial files and folders
