+++ lib/kits/paged-publication/index.js
@@ -1,5 +1,3 @@
 export {default as Viewer} from './viewer';
 
 export {default as Bootstrapper} from './bootstrapper';
-
-export {default as TemplateBuilder} from './template-builder';
