+++ lib/kits/templates/components/index.js
