+++ lib/kits/templates/components/main-container.js
@@ -1,5 +1,5 @@
-// import './main-container.styl';
 import Mustache from 'mustache';
+import './main-container.styl';
 
 const defaultTemplate = `\
     <div class="sgn__pp" data-layout-absolute="true">
