+++ lib/kits/templates/components/main-container.styl
