+++ lib/kits/templates/index.js
@@ -1,0 +1,1 @@
+export {default as PagedPublication} from './paged-publication';
