+++ lib/kits/templates/paged-publication.js
@@ -1,8 +1,8 @@
-import Bootstrapper from './bootstrapper';
+import * as IncitoPublicationKit from '../incito-publication';
+import * as PagedPublicationKit from '../paged-publication';
+import * as components from './components';
 import {request} from '../core';
-import * as components from '../core-ui/components';
 import SGN from '../../sgn-sdk';
-import './template-builder.styl';
 
 export default class TemplateBuilder {
     constructor(scriptEl) {
@@ -62,7 +62,7 @@ export default class TemplateBuilder {
     }
 
     start() {
-        const bootstrapper = new Bootstrapper(this.options);
+        const bootstrapper = new PagedPublicationKit.Bootstrapper(this.options);
 
         bootstrapper.fetch((err, data) => {
             if (!err) {
