+++ lib/sgn-sdk.js
@@ -4,6 +4,7 @@ import * as CoreUIKit from './kits/core-ui';
 import * as EventsKit from './kits/events';
 import * as IncitoPublicationKit from './kits/incito-publication';
 import * as PagedPublicationKit from './kits/paged-publication';
+import * as Templates from './kits/templates';
 import * as clientLocal from './storage/client-local';
 import './stylus/sgn.styl';
 import * as translations from './translations';
@@ -25,6 +26,7 @@ SGN.CoreKit = CoreKit;
 SGN.PagedPublicationKit = PagedPublicationKit;
 SGN.IncitoPublicationKit = IncitoPublicationKit;
 SGN.CoreUIKit = CoreUIKit;
+SGN.Templates = Templates;
 
 SGN.config.bind('change', (changedAttributes) => {
     const newEventTracker = changedAttributes.eventTracker;
@@ -87,7 +89,7 @@ if (isBrowser()) {
         SGN.config.set(config);
 
         if (pubController === 'view-publication') {
-            const templateBuilder = new SGN.PagedPublicationKit.TemplateBuilder(
+            const templateBuilder = new SGN.Templates.PagedPublication(
                 scriptEl
             );
